COMP: Fixed warning.

+++ VTK/Parallel/vtkMultiProcessController.cxx
@@ -456,7 +456,8 @@ int vtkMultiProcessController::ProcessRMIs(int reportErrors, int dont_loop)
       // use the inline data or make a second receive to fetch the data.
       if (static_cast<unsigned int>(triggerMessage[1]) < sizeof(int)*(128-4))
         {
-        int num_ints = 4 + ceil(triggerMessage[1]/static_cast<double>(sizeof(int)));
+        int num_ints = 4 + static_cast<int>(
+            ceil(triggerMessage[1]/static_cast<double>(sizeof(int))));
         if (this->RMICommunicator->GetCount() != num_ints)
           {
           if (reportErrors)
